Control Settings for Stochastic Individual Contact Models

Sets the controls for stochastic individual contact models simulated with icm().

Usage

control.icm(type, nsteps, nsims = 1, verbose = FALSE, verbose.int = 0)

Arguments

type

Disease type to be modeled, with the choice of "SI" for Susceptible-Infected diseases, "SIR" for Susceptible-Infected-Recovered diseases, and "SIS" for Susceptible-Infected-Susceptible diseases.

nsteps

Number of time steps to solve the model over. This must be a positive integer.

nsims

Number of simulations to run.

verbose

If TRUE, print model progress to the console.

verbose.int

Time step interval for printing progress to console, where 0 (the default) prints completion status of entire simulation and positive integer x prints progress after every x time steps.

Value

An EpiModel object of class control.icm.

Details

control.icm sets the required control settings for any stochastic individual contact model solved with the icm() function. ICM simulations use the built-in SI, SIR, and SIS disease types only. The type argument is required and has no default. For custom or extension epidemic models, use the network model class via control.net() instead.

See also