EpiModel Web

Runs a web browser-based GUI of deterministic compartmental models, stochastic individual contact models, and basic network models.

Usage

epiweb(class, ...)

Arguments

class

Model class, with options of "dcm", "icm", and "net".

...

Additional arguments passed to shiny::runApp.

Details

epiweb runs a web-based GUI of one-group deterministic compartmental models, stochastic individual contact models, and stochastic network models with user input on model type, state sizes, and parameters. Model output may be plotted, summarized, and saved as raw data using the core EpiModel functionality for these model classes. These applications are built using the shiny package framework.

References

RStudio. shiny: Web Application Framework for R. R package version 1.0.5. 2015. https://shiny.posit.co/.

See also

Examples

if (FALSE) {
## Deterministic compartmental models
epiweb(class = "dcm")

## Stochastic individual contact models
epiweb(class = "icm")

## Stochastic network models
epiweb(class = "net")
}