Extract Model Data for Stochastic Models

This function extracts model simulations for objects of classes icm and netsim into a data frame using the generic as.data.frame function.

Usage

# S3 method for icm
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  out = "vals",
  sim,
  qval,
  ...
)

# S3 method for netsim
as.data.frame(x, row.names = NULL, optional = FALSE, out = "vals", sim, ...)

Arguments

x

An EpiModel object of class icm or netsim.

row.names

See as.data.frame.default.

optional

See as.data.frame.default.

out

Data output to data frame: "mean" for row means across simulations, "sd" for row standard deviations across simulations, "qnt" for row quantiles at the level specified in qval, or "vals" for values from individual simulations.

sim

If out="vals", the simulation number to output. If not specified, then data from all simulations will be output.

qval

Quantile value required when out="qnt".

...

See as.data.frame.default.

Value

A data frame containing the data from x.

Details

These methods work for both icm and netsim class models. The available output includes time-specific means, standard deviations, quantiles, and simulation values (compartment and flow sizes) from these stochastic model classes. Means, standard deviations, and quantiles are calculated by taking the row summary (i.e., each row of data is corresponds to a time step) across all simulations in the model output.

Examples

## Stochastic ICM SIS model
param <- param.icm(inf.prob = 0.8, act.rate = 2, rec.rate = 0.1)
init <- init.icm(s.num = 500, i.num = 1)
control <- control.icm(type = "SIS", nsteps = 10,
                       nsims = 3, verbose = FALSE)
mod <- icm(param, init, control)

# Default output all simulation runs, default to all in stacked data.frame
as.data.frame(mod)
#>    sim time s.num i.num num si.flow is.flow
#> 1    1    1   500     1 501       0       0
#> 2    1    2   500     1 501       0       0
#> 3    1    3   498     3 501       2       0
#> 4    1    4   497     4 501       1       0
#> 5    1    5   496     5 501       4       3
#> 6    1    6   488    13 501       8       0
#> 7    1    7   472    29 501      20       4
#> 8    1    8   438    63 501      41       7
#> 9    1    9   370   131 501      84      16
#> 10   1   10   276   225 501     125      31
#> 11   2    1   500     1 501       0       0
#> 12   2    2   499     2 501       1       0
#> 13   2    3   497     4 501       2       0
#> 14   2    4   496     5 501       3       2
#> 15   2    5   489    12 501      11       4
#> 16   2    6   470    31 501      20       1
#> 17   2    7   440    61 501      37       7
#> 18   2    8   366   135 501      90      16
#> 19   2    9   273   228 501     120      27
#> 20   2   10   169   332 501     140      36
#> 21   3    1   500     1 501       0       0
#> 22   3    2   498     3 501       2       0
#> 23   3    3   494     7 501       6       2
#> 24   3    4   488    13 501       8       2
#> 25   3    5   471    30 501      17       0
#> 26   3    6   423    78 501      52       4
#> 27   3    7   343   158 501      93      13
#> 28   3    8   244   257 501     131      32
#> 29   3    9   133   368 501     149      38
#> 30   3   10    82   419 501      95      44
as.data.frame(mod, sim = 2)
#>    sim time s.num i.num num si.flow is.flow
#> 1    2    1   500     1 501       0       0
#> 2    2    2   499     2 501       1       0
#> 3    2    3   497     4 501       2       0
#> 4    2    4   496     5 501       3       2
#> 5    2    5   489    12 501      11       4
#> 6    2    6   470    31 501      20       1
#> 7    2    7   440    61 501      37       7
#> 8    2    8   366   135 501      90      16
#> 9    2    9   273   228 501     120      27
#> 10   2   10   169   332 501     140      36

# Time-specific means across simulations
as.data.frame(mod, out = "mean")
#>    time    s.num      i.num num    si.flow    is.flow
#> 1     1 500.0000   1.000000 501   0.000000  0.0000000
#> 2     2 499.0000   2.000000 501   1.000000  0.0000000
#> 3     3 496.3333   4.666667 501   3.333333  0.6666667
#> 4     4 493.6667   7.333333 501   4.000000  1.3333333
#> 5     5 485.3333  15.666667 501  10.666667  2.3333333
#> 6     6 460.3333  40.666667 501  26.666667  1.6666667
#> 7     7 418.3333  82.666667 501  50.000000  8.0000000
#> 8     8 349.3333 151.666667 501  87.333333 18.3333333
#> 9     9 258.6667 242.333333 501 117.666667 27.0000000
#> 10   10 175.6667 325.333333 501 120.000000 37.0000000

# Time-specific standard deviations across simulations
as.data.frame(mod, out = "sd")
#>    time      s.num      i.num num   si.flow   is.flow
#> 1     1   0.000000   0.000000   0  0.000000  0.000000
#> 2     2   1.000000   1.000000   0  1.000000  0.000000
#> 3     3   2.081666   2.081666   0  2.309401  1.154701
#> 4     4   4.932883   4.932883   0  3.605551  1.154701
#> 5     5  12.897028  12.897028   0  6.506407  2.081666
#> 6     6  33.560890  33.560890   0 22.744963  2.081666
#> 7     7  67.173904  67.173904   0 38.196859  4.582576
#> 8     8  98.068004  98.068004   0 45.059220 12.662280
#> 9     9 119.148367 119.148367   0 32.562760 11.000000
#> 10   10  97.171669  97.171669   0 22.912878  6.557439

# Time-specific quantile values across simulations
as.data.frame(mod, out = "qnt", qval = 0.25)
#>    time s.num i.num num si.flow is.flow
#> 1     1 500.0   1.0 501     0.0     0.0
#> 2     2 498.5   1.5 501     0.5     0.0
#> 3     3 495.5   3.5 501     2.0     0.0
#> 4     4 492.0   4.5 501     2.0     1.0
#> 5     5 480.0   8.5 501     7.5     1.5
#> 6     6 446.5  22.0 501    14.0     0.5
#> 7     7 391.5  45.0 501    28.5     5.5
#> 8     8 305.0  99.0 501    65.5    11.5
#> 9     9 203.0 179.5 501   102.0    21.5
#> 10   10 125.5 278.5 501   110.0    33.5
as.data.frame(mod, out = "qnt", qval = 0.75)
#>    time s.num i.num num si.flow is.flow
#> 1     1 500.0   1.0 501     0.0     0.0
#> 2     2 499.5   2.5 501     1.5     0.0
#> 3     3 497.5   5.5 501     4.0     1.0
#> 4     4 496.5   9.0 501     5.5     2.0
#> 5     5 492.5  21.0 501    14.0     3.5
#> 6     6 479.0  54.5 501    36.0     2.5
#> 7     7 456.0 109.5 501    65.0    10.0
#> 8     8 402.0 196.0 501   110.5    24.0
#> 9     9 321.5 298.0 501   134.5    32.5
#> 10   10 222.5 375.5 501   132.5    40.0

if (FALSE) {
## Stochastic SI Network Model
nw <- network_initialize(n = 100)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)
est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)

param <- param.net(inf.prob = 0.5)
init <- init.net(i.num = 10)
control <- control.net(type = "SI", nsteps = 10, nsims = 3, verbose = FALSE)
mod <- netsim(est, param, init, control)

# Same data extraction methods as with ICMs
as.data.frame(mod)
as.data.frame(mod, sim = 2)
as.data.frame(mod, out = "mean")
as.data.frame(mod, out = "sd")
as.data.frame(mod, out = "qnt", qval = 0.25)
as.data.frame(mod, out = "qnt", qval = 0.75)
}