Extract Timed Edgelists for netdx Objects

This function extracts timed edgelists for objects of class netdx into a data frame using the generic as.data.frame function.

Usage

# S3 method for netdx
as.data.frame(x, row.names = NULL, optional = FALSE, sim, ...)

Arguments

x

An EpiModel object of class netdx.

row.names

See as.data.frame.default.

optional

See as.data.frame.default.

sim

The simulation number to output. If not specified, then data from all simulations will be output.

...

See as.data.frame.default.

Value

A data frame containing the data from x.

Examples

# Initialize and parameterize the network model
nw <- network_initialize(n = 100)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)

# Model estimation
est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)
#> Starting maximum pseudolikelihood estimation (MPLE):
#> Obtaining the responsible dyads.
#> Evaluating the predictor and response matrix.
#> Maximizing the pseudolikelihood.
#> Finished MPLE.

# Simulate the network with netdx
dx <- netdx(est, nsims = 3, nsteps = 10, keep.tedgelist = TRUE,
            verbose = FALSE)

# Extract data from the first simulation
as.data.frame(dx, sim = 1)
#>    onset terminus tail head onset.censored terminus.censored duration edge.id
#> 1      0       11    1   57          FALSE              TRUE       11       1
#> 2      0       11    1   75          FALSE              TRUE       11       2
#> 3      0       11    4    9          FALSE              TRUE       11       3
#> 4      0       11    4   83          FALSE              TRUE       11       4
#> 5      0        1    6   13          FALSE             FALSE        1       5
#> 6      0        7    7   54          FALSE             FALSE        7       6
#> 7      0        4    9   47          FALSE             FALSE        4       7
#> 8      0       11   10   71          FALSE              TRUE       11       8
#> 9      0        8   10   78          FALSE             FALSE        8       9
#> 10     0       11   11   30          FALSE              TRUE       11      10
#> 11     0        4   11   67          FALSE             FALSE        4      11
#> 12     0        8   11   68          FALSE             FALSE        8      12
#> 13     0       11   11   82          FALSE              TRUE       11      13
#> 14     0       11   12   33          FALSE              TRUE       11      14
#> 15     0        2   12   56          FALSE             FALSE        2      15
#> 16     0       11   13   98          FALSE              TRUE       11      16
#> 17     0        5   17   23          FALSE             FALSE        5      17
#> 18     0       11   18   79          FALSE              TRUE       11      18
#> 19     0        3   26   43          FALSE             FALSE        3      19
#> 20     0       11   27   35          FALSE              TRUE       11      20
#> 21     0        6   27  100          FALSE             FALSE        6      21
#> 22     0       10   29   66          FALSE             FALSE       10      22
#> 23     0       11   32   52          FALSE              TRUE       11      23
#> 24     0        5   36   40          FALSE             FALSE        5      24
#> 25     0       11   37   65          FALSE              TRUE       11      25
#> 26     0       11   40   74          FALSE              TRUE       11      26
#> 27     0       11   40   88          FALSE              TRUE       11      27
#> 28     0        1   44   98          FALSE             FALSE        1      28
#> 29     0       11   50   51          FALSE              TRUE       11      29
#> 30     0       11   50   61          FALSE              TRUE       11      30
#> 31     0       11   51   75          FALSE              TRUE       11      31
#> 32     0       11   53   86          FALSE              TRUE       11      32
#> 33     0       11   54   80          FALSE              TRUE       11      33
#> 34     0       10   55   59          FALSE             FALSE       10      34
#> 35     0       11   55   68          FALSE              TRUE       11      35
#> 36     0        2   57   99          FALSE             FALSE        2      36
#> 37     0        8   65   70          FALSE             FALSE        8      37
#> 38     0       11   72   91          FALSE              TRUE       11      38
#> 39     1       11   20   92          FALSE              TRUE       10      39
#> 40     1       11    7   53          FALSE              TRUE       10      40
#> 41     1       11    7   52          FALSE              TRUE       10      41
#> 42     1       11   11   55          FALSE              TRUE       10      42
#> 43     3        5    5   30          FALSE             FALSE        2      43
#> 44     3       11   28   84          FALSE              TRUE        8      44
#> 45     4       11   42   60          FALSE              TRUE        7      45
#> 46     5       10   60   70          FALSE             FALSE        5      46
#> 47     5       11    7   80          FALSE              TRUE        6      47
#> 48     5       11    2   76          FALSE              TRUE        6      48
#> 49     5       11    9   66          FALSE              TRUE        6      49
#> 50     6       11   30   51          FALSE              TRUE        5      50
#> 51     6       11   58   62          FALSE              TRUE        5      51
#> 52     6       11   22  100          FALSE              TRUE        5      52
#> 53     6       11   17   52          FALSE              TRUE        5      53
#> 54     6       11   13   18          FALSE              TRUE        5      54
#> 55     7       11   72   73          FALSE              TRUE        4      55
#> 56     8       11   37   42          FALSE              TRUE        3      56
#> 57     8       10   49   83          FALSE             FALSE        2      57
#> 58     8       11   58  100          FALSE              TRUE        3      58
#> 59     9       11   73   88          FALSE              TRUE        2      59
#> 60     9       11   56   81          FALSE              TRUE        2      60
#> 61    10       11   34   40          FALSE              TRUE        1      61

# Extract data from all simulations
as.data.frame(dx)
#>     sim onset terminus tail head onset.censored terminus.censored duration
#> 1     1     0       11    1   57          FALSE              TRUE       11
#> 2     1     0       11    1   75          FALSE              TRUE       11
#> 3     1     0       11    4    9          FALSE              TRUE       11
#> 4     1     0       11    4   83          FALSE              TRUE       11
#> 5     1     0        1    6   13          FALSE             FALSE        1
#> 6     1     0        7    7   54          FALSE             FALSE        7
#> 7     1     0        4    9   47          FALSE             FALSE        4
#> 8     1     0       11   10   71          FALSE              TRUE       11
#> 9     1     0        8   10   78          FALSE             FALSE        8
#> 10    1     0       11   11   30          FALSE              TRUE       11
#> 11    1     0        4   11   67          FALSE             FALSE        4
#> 12    1     0        8   11   68          FALSE             FALSE        8
#> 13    1     0       11   11   82          FALSE              TRUE       11
#> 14    1     0       11   12   33          FALSE              TRUE       11
#> 15    1     0        2   12   56          FALSE             FALSE        2
#> 16    1     0       11   13   98          FALSE              TRUE       11
#> 17    1     0        5   17   23          FALSE             FALSE        5
#> 18    1     0       11   18   79          FALSE              TRUE       11
#> 19    1     0        3   26   43          FALSE             FALSE        3
#> 20    1     0       11   27   35          FALSE              TRUE       11
#> 21    1     0        6   27  100          FALSE             FALSE        6
#> 22    1     0       10   29   66          FALSE             FALSE       10
#> 23    1     0       11   32   52          FALSE              TRUE       11
#> 24    1     0        5   36   40          FALSE             FALSE        5
#> 25    1     0       11   37   65          FALSE              TRUE       11
#> 26    1     0       11   40   74          FALSE              TRUE       11
#> 27    1     0       11   40   88          FALSE              TRUE       11
#> 28    1     0        1   44   98          FALSE             FALSE        1
#> 29    1     0       11   50   51          FALSE              TRUE       11
#> 30    1     0       11   50   61          FALSE              TRUE       11
#> 31    1     0       11   51   75          FALSE              TRUE       11
#> 32    1     0       11   53   86          FALSE              TRUE       11
#> 33    1     0       11   54   80          FALSE              TRUE       11
#> 34    1     0       10   55   59          FALSE             FALSE       10
#> 35    1     0       11   55   68          FALSE              TRUE       11
#> 36    1     0        2   57   99          FALSE             FALSE        2
#> 37    1     0        8   65   70          FALSE             FALSE        8
#> 38    1     0       11   72   91          FALSE              TRUE       11
#> 39    1     1       11   20   92          FALSE              TRUE       10
#> 40    1     1       11    7   53          FALSE              TRUE       10
#> 41    1     1       11    7   52          FALSE              TRUE       10
#> 42    1     1       11   11   55          FALSE              TRUE       10
#> 43    1     3        5    5   30          FALSE             FALSE        2
#> 44    1     3       11   28   84          FALSE              TRUE        8
#> 45    1     4       11   42   60          FALSE              TRUE        7
#> 46    1     5       10   60   70          FALSE             FALSE        5
#> 47    1     5       11    7   80          FALSE              TRUE        6
#> 48    1     5       11    2   76          FALSE              TRUE        6
#> 49    1     5       11    9   66          FALSE              TRUE        6
#> 50    1     6       11   30   51          FALSE              TRUE        5
#> 51    1     6       11   58   62          FALSE              TRUE        5
#> 52    1     6       11   22  100          FALSE              TRUE        5
#> 53    1     6       11   17   52          FALSE              TRUE        5
#> 54    1     6       11   13   18          FALSE              TRUE        5
#> 55    1     7       11   72   73          FALSE              TRUE        4
#> 56    1     8       11   37   42          FALSE              TRUE        3
#> 57    1     8       10   49   83          FALSE             FALSE        2
#> 58    1     8       11   58  100          FALSE              TRUE        3
#> 59    1     9       11   73   88          FALSE              TRUE        2
#> 60    1     9       11   56   81          FALSE              TRUE        2
#> 61    1    10       11   34   40          FALSE              TRUE        1
#> 62    2     0       11    2   34          FALSE              TRUE       11
#> 63    2     0       11    4   11          FALSE              TRUE       11
#> 64    2     0       11    4   33          FALSE              TRUE       11
#> 65    2     0       11    4   83          FALSE              TRUE       11
#> 66    2     0       11    5   35          FALSE              TRUE       11
#> 67    2     0       11    5   64          FALSE              TRUE       11
#> 68    2     0       11    5   97          FALSE              TRUE       11
#> 69    2     0       11    6   63          FALSE              TRUE       11
#> 70    2     0       11    7   74          FALSE              TRUE       11
#> 71    2     0       11    8   45          FALSE              TRUE       11
#> 72    2     0       11    8   63          FALSE              TRUE       11
#> 73    2     0       11   10   17          FALSE              TRUE       11
#> 74    2     0       11   10   70          FALSE              TRUE       11
#> 75    2     0       11   12   40          FALSE              TRUE       11
#> 76    2     0       11   13   61          FALSE              TRUE       11
#> 77    2     0        8   14   89          FALSE             FALSE        8
#> 78    2     0       11   15   24          FALSE              TRUE       11
#> 79    2     0        3   16   78          FALSE             FALSE        3
#> 80    2     0       11   19   36          FALSE              TRUE       11
#> 81    2     0       11   22   32          FALSE              TRUE       11
#> 82    2     0       11   22   36          FALSE              TRUE       11
#> 83    2     0        3   23   31          FALSE             FALSE        3
#> 84    2     0       11   23   67          FALSE              TRUE       11
#> 85    2     0       11   24   39          FALSE              TRUE       11
#> 86    2     0       11   24   48          FALSE              TRUE       11
#> 87    2     0        5   24   95          FALSE             FALSE        5
#> 88    2     0        3   26   56          FALSE             FALSE        3
#> 89    2     0       11   27   64          FALSE              TRUE       11
#> 90    2     0        9   29   43          FALSE             FALSE        9
#> 91    2     0       11   30   50          FALSE              TRUE       11
#> 92    2     0        6   30   64          FALSE             FALSE        6
#> 93    2     0       11   35   68          FALSE              TRUE       11
#> 94    2     0       11   35   72          FALSE              TRUE       11
#> 95    2     0       11   37   90          FALSE              TRUE       11
#> 96    2     0        7   38   45          FALSE             FALSE        7
#> 97    2     0        1   38  100          FALSE             FALSE        1
#> 98    2     0       11   40   80          FALSE              TRUE       11
#> 99    2     0       11   42   81          FALSE              TRUE       11
#> 100   2     0       11   47  100          FALSE              TRUE       11
#> 101   2     0       11   52   83          FALSE              TRUE       11
#> 102   2     0       11   55   98          FALSE              TRUE       11
#> 103   2     0        8   55   99          FALSE             FALSE        8
#> 104   2     0        2   56   63          FALSE             FALSE        2
#> 105   2     0       11   58   66          FALSE              TRUE       11
#> 106   2     0       11   59   71          FALSE              TRUE       11
#> 107   2     0        1   60   89          FALSE             FALSE        1
#> 108   2     0       11   63   81          FALSE              TRUE       11
#> 109   2     0        3   66   69          FALSE             FALSE        3
#> 110   2     0       11   66   95          FALSE              TRUE       11
#> 111   2     0       11   73   99          FALSE              TRUE       11
#> 112   2     0       10   74   82          FALSE             FALSE       10
#> 113   2     0       11   74   94          FALSE              TRUE       11
#> 114   2     0       11   75   82          FALSE              TRUE       11
#> 115   2     0       11   77   90          FALSE              TRUE       11
#> 116   2     0       11   90   99          FALSE              TRUE       11
#> 117   2     1        6   40   56          FALSE             FALSE        5
#> 118   2     1       11    4   29          FALSE              TRUE       10
#> 119   2     1        3   10   89          FALSE             FALSE        2
#> 120   2     1        7   27   83          FALSE             FALSE        6
#> 121   2     1       11   38   78          FALSE              TRUE       10
#> 122   2     1        8   43   87          FALSE             FALSE        7
#> 123   2     2       11   32   99          FALSE              TRUE        9
#> 124   2     2       11   62   65          FALSE              TRUE        9
#> 125   2     2       11   18   41          FALSE              TRUE        9
#> 126   2     2        4   18   81          FALSE             FALSE        2
#> 127   2     2        3   46   55          FALSE             FALSE        1
#> 128   2     3       10   10   60          FALSE             FALSE        7
#> 129   2     3        7   27   60          FALSE             FALSE        4
#> 130   2     3       11   21   41          FALSE              TRUE        8
#> 131   2     3       11   29   78          FALSE              TRUE        8
#> 132   2     4       11   30   31          FALSE              TRUE        7
#> 133   2     4       10   19   29          FALSE             FALSE        6
#> 134   2     4       11    7   81          FALSE              TRUE        7
#> 135   2     5       11    8   54          FALSE              TRUE        6
#> 136   2     5       11   16   63          FALSE              TRUE        6
#> 137   2     5       10   59   68          FALSE             FALSE        5
#> 138   2     5       11   67   75          FALSE              TRUE        6
#> 139   2     5       11   50   87          FALSE              TRUE        6
#> 140   2     6       11   60   84          FALSE              TRUE        5
#> 141   2     6       11   26   60          FALSE              TRUE        5
#> 142   2     6       11   52   87          FALSE              TRUE        5
#> 143   2     8       11    4   65          FALSE              TRUE        3
#> 144   2     9       11    7   71          FALSE              TRUE        2
#> 145   2     9       11    7   61          FALSE              TRUE        2
#> 146   2     9       11   27   96          FALSE              TRUE        2
#> 147   2    10       11   65   90          FALSE              TRUE        1
#> 148   2    10       11    1   46          FALSE              TRUE        1
#> 149   2    10       11   28   54          FALSE              TRUE        1
#> 150   2    10       11   77   99          FALSE              TRUE        1
#> 151   2    10       11   38   86          FALSE              TRUE        1
#> 152   2    10       11   18   98          FALSE              TRUE        1
#> 153   2    10       11   38   73          FALSE              TRUE        1
#> 154   3     0       11    1   38          FALSE              TRUE       11
#> 155   3     0       11    2   41          FALSE              TRUE       11
#> 156   3     0        3    3   35          FALSE             FALSE        3
#> 157   3     0       11    5    7          FALSE              TRUE       11
#> 158   3     0       11    7    9          FALSE              TRUE       11
#> 159   3     0        9    7   83          FALSE             FALSE        9
#> 160   3     0        3    8   87          FALSE             FALSE        3
#> 161   3     0       11   10   22          FALSE              TRUE       11
#> 162   3     0       11   11   18          FALSE              TRUE       11
#> 163   3     0       11   11   45          FALSE              TRUE       11
#> 164   3     0       11   12   27          FALSE              TRUE       11
#> 165   3     0        9   14   78          FALSE             FALSE        9
#> 166   3     0       11   15   73          FALSE              TRUE       11
#> 167   3     0       11   16   81          FALSE              TRUE       11
#> 168   3     0       11   18   36          FALSE              TRUE       11
#> 169   3     0       11   18   81          FALSE              TRUE       11
#> 170   3     0       11   21   72          FALSE              TRUE       11
#> 171   3     0       11   21   74          FALSE              TRUE       11
#> 172   3     0        9   22   56          FALSE             FALSE        9
#> 173   3     0       11   22   90          FALSE              TRUE       11
#> 174   3     0        5   23   70          FALSE             FALSE        5
#> 175   3     0        1   25   82          FALSE             FALSE        1
#> 176   3     0       10   26   85          FALSE             FALSE       10
#> 177   3     0       11   26   92          FALSE              TRUE       11
#> 178   3     0        1   27   34          FALSE             FALSE        1
#> 179   3     0       11   27   61          FALSE              TRUE       11
#> 180   3     0       11   27   72          FALSE              TRUE       11
#> 181   3     0        8   28   88          FALSE             FALSE        8
#> 182   3     0        1   29   98          FALSE             FALSE        1
#> 183   3     0        8   30   60          FALSE             FALSE        8
#> 184   3     0       11   32   49          FALSE              TRUE       11
#> 185   3     0       11   34   64          FALSE              TRUE       11
#> 186   3     0        4   35   80          FALSE             FALSE        4
#> 187   3     0       11   36   67          FALSE              TRUE       11
#> 188   3     0        9   38   74          FALSE             FALSE        9
#> 189   3     0       11   39   86          FALSE              TRUE       11
#> 190   3     0        7   40   57          FALSE             FALSE        7
#> 191   3     0       11   43   94          FALSE              TRUE       11
#> 192   3     0        1   44   87          FALSE             FALSE        1
#> 193   3     0        1   46   86          FALSE             FALSE        1
#> 194   3     0       11   48   80          FALSE              TRUE       11
#> 195   3     0        5   49   83          FALSE             FALSE        5
#> 196   3     0       11   51   71          FALSE              TRUE       11
#> 197   3     0       11   51   89          FALSE              TRUE       11
#> 198   3     0       11   55   71          FALSE              TRUE       11
#> 199   3     0       11   56   67          FALSE              TRUE       11
#> 200   3     0       10   59   95          FALSE             FALSE       10
#> 201   3     0       11   61   94          FALSE              TRUE       11
#> 202   3     0        5   62   75          FALSE             FALSE        5
#> 203   3     0        1   63   96          FALSE             FALSE        1
#> 204   3     0        5   68   93          FALSE             FALSE        5
#> 205   3     0       10   69   72          FALSE             FALSE       10
#> 206   3     0       11   70   80          FALSE              TRUE       11
#> 207   3     0        6   70   95          FALSE             FALSE        6
#> 208   3     0       11   73   88          FALSE              TRUE       11
#> 209   3     0       11   79   99          FALSE              TRUE       11
#> 210   3     0       11   81  100          FALSE              TRUE       11
#> 211   3     0       11   83   92          FALSE              TRUE       11
#> 212   3     0       11   86   89          FALSE              TRUE       11
#> 213   3     0        5   90   95          FALSE             FALSE        5
#> 214   3     1       11   43   82          FALSE              TRUE       10
#> 215   3     1        3   31   97          FALSE             FALSE        2
#> 216   3     1       11   67   97          FALSE              TRUE       10
#> 217   3     1        8   56   86          FALSE             FALSE        7
#> 218   3     2        4   35   93          FALSE             FALSE        2
#> 219   3     3       11   61   87          FALSE              TRUE        8
#> 220   3     3       11   22   94          FALSE              TRUE        8
#> 221   3     3        6    4   21          FALSE             FALSE        3
#> 222   3     3       11   44   81          FALSE              TRUE        8
#> 223   3     3       10    1   14          FALSE             FALSE        7
#> 224   3     4        5   48   81          FALSE             FALSE        1
#> 225   3     4        5    2   11          FALSE             FALSE        1
#> 226   3     5        6    9   54          FALSE             FALSE        1
#> 227   3     5       11   16   83          FALSE              TRUE        6
#> 228   3     5       11   44   72          FALSE              TRUE        6
#> 229   3     6       11   18   87          FALSE              TRUE        5
#> 230   3     7       10   42   76          FALSE             FALSE        3
#> 231   3     7        9   17   66          FALSE             FALSE        2
#> 232   3     7       11   48   98          FALSE              TRUE        4
#> 233   3     9       11   18   35          FALSE              TRUE        2
#> 234   3     9       11   10   24          FALSE              TRUE        2
#> 235   3    10       11   33   34          FALSE              TRUE        1
#> 236   3    10       11   55   80          FALSE              TRUE        1
#> 237   3    10       11    8   47          FALSE              TRUE        1
#> 238   3    10       11   50   61          FALSE              TRUE        1
#> 239   3    10       11   26   56          FALSE              TRUE        1
#>     edge.id
#> 1         1
#> 2         2
#> 3         3
#> 4         4
#> 5         5
#> 6         6
#> 7         7
#> 8         8
#> 9         9
#> 10       10
#> 11       11
#> 12       12
#> 13       13
#> 14       14
#> 15       15
#> 16       16
#> 17       17
#> 18       18
#> 19       19
#> 20       20
#> 21       21
#> 22       22
#> 23       23
#> 24       24
#> 25       25
#> 26       26
#> 27       27
#> 28       28
#> 29       29
#> 30       30
#> 31       31
#> 32       32
#> 33       33
#> 34       34
#> 35       35
#> 36       36
#> 37       37
#> 38       38
#> 39       39
#> 40       40
#> 41       41
#> 42       42
#> 43       43
#> 44       44
#> 45       45
#> 46       46
#> 47       47
#> 48       48
#> 49       49
#> 50       50
#> 51       51
#> 52       52
#> 53       53
#> 54       54
#> 55       55
#> 56       56
#> 57       57
#> 58       58
#> 59       59
#> 60       60
#> 61       61
#> 62        1
#> 63        2
#> 64        3
#> 65        4
#> 66        5
#> 67        6
#> 68        7
#> 69        8
#> 70        9
#> 71       10
#> 72       11
#> 73       12
#> 74       13
#> 75       14
#> 76       15
#> 77       16
#> 78       17
#> 79       18
#> 80       19
#> 81       20
#> 82       21
#> 83       22
#> 84       23
#> 85       24
#> 86       25
#> 87       26
#> 88       27
#> 89       28
#> 90       29
#> 91       30
#> 92       31
#> 93       32
#> 94       33
#> 95       34
#> 96       35
#> 97       36
#> 98       37
#> 99       38
#> 100      39
#> 101      40
#> 102      41
#> 103      42
#> 104      43
#> 105      44
#> 106      45
#> 107      46
#> 108      47
#> 109      48
#> 110      49
#> 111      50
#> 112      51
#> 113      52
#> 114      53
#> 115      54
#> 116      55
#> 117      56
#> 118      57
#> 119      58
#> 120      59
#> 121      60
#> 122      61
#> 123      62
#> 124      63
#> 125      64
#> 126      65
#> 127      66
#> 128      67
#> 129      68
#> 130      69
#> 131      70
#> 132      71
#> 133      72
#> 134      73
#> 135      74
#> 136      75
#> 137      76
#> 138      77
#> 139      78
#> 140      79
#> 141      80
#> 142      81
#> 143      82
#> 144      83
#> 145      84
#> 146      85
#> 147      86
#> 148      87
#> 149      88
#> 150      89
#> 151      90
#> 152      91
#> 153      92
#> 154       1
#> 155       2
#> 156       3
#> 157       4
#> 158       5
#> 159       6
#> 160       7
#> 161       8
#> 162       9
#> 163      10
#> 164      11
#> 165      12
#> 166      13
#> 167      14
#> 168      15
#> 169      16
#> 170      17
#> 171      18
#> 172      19
#> 173      20
#> 174      21
#> 175      22
#> 176      23
#> 177      24
#> 178      25
#> 179      26
#> 180      27
#> 181      28
#> 182      29
#> 183      30
#> 184      31
#> 185      32
#> 186      33
#> 187      34
#> 188      35
#> 189      36
#> 190      37
#> 191      38
#> 192      39
#> 193      40
#> 194      41
#> 195      42
#> 196      43
#> 197      44
#> 198      45
#> 199      46
#> 200      47
#> 201      48
#> 202      49
#> 203      50
#> 204      51
#> 205      52
#> 206      53
#> 207      54
#> 208      55
#> 209      56
#> 210      57
#> 211      58
#> 212      59
#> 213      60
#> 214      61
#> 215      62
#> 216      63
#> 217      64
#> 218      65
#> 219      66
#> 220      67
#> 221      68
#> 222      69
#> 223      70
#> 224      71
#> 225      72
#> 226      73
#> 227      74
#> 228      75
#> 229      76
#> 230      77
#> 231      78
#> 232      79
#> 233      80
#> 234      81
#> 235      82
#> 236      83
#> 237      84
#> 238      85
#> 239      86