Extract Timed Edgelists for netdx Objects

This function extracts timed edgelists for objects of class netdx into a data frame using the generic as.data.frame function.

Usage

# S3 method for netdx
as.data.frame(x, row.names = NULL, optional = FALSE, sim, ...)

Arguments

x

An EpiModel object of class netdx.

row.names

See as.data.frame.default.

optional

See as.data.frame.default.

sim

The simulation number to output. If not specified, then data from all simulations will be output.

...

See as.data.frame.default.

Value

A data frame containing the data from x.

Examples

# Initialize and parameterize the network model
nw <- network_initialize(n = 100)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)

# Model estimation
est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)
#> Starting maximum pseudolikelihood estimation (MPLE):
#> Obtaining the responsible dyads.
#> Evaluating the predictor and response matrix.
#> Maximizing the pseudolikelihood.
#> Finished MPLE.

# Simulate the network with netdx
dx <- netdx(est, nsims = 3, nsteps = 10, keep.tedgelist = TRUE,
            verbose = FALSE)

# Extract data from the first simulation
as.data.frame(dx, sim = 1)
#>    onset terminus tail head onset.censored terminus.censored duration edge.id
#> 1      0       11    2   67          FALSE              TRUE       11       1
#> 2      0       11    3   62          FALSE              TRUE       11       2
#> 3      0       11    4   16          FALSE              TRUE       11       3
#> 4      0       11    4   58          FALSE              TRUE       11       4
#> 5      0       11    5   39          FALSE              TRUE       11       5
#> 6      0        7    6   61          FALSE             FALSE        7       6
#> 7      0        5    6   70          FALSE             FALSE        5       7
#> 8      0       11    6   79          FALSE              TRUE       11       8
#> 9      0       11    7   74          FALSE              TRUE       11       9
#> 10     0        1    8   50          FALSE             FALSE        1      10
#> 11     0        3   10   79          FALSE             FALSE        3      11
#> 12     0       11   10   81          FALSE              TRUE       11      12
#> 13     0       11   10   86          FALSE              TRUE       11      13
#> 14     0       10   11   55          FALSE             FALSE       10      14
#> 15     0       11   11   85          FALSE              TRUE       11      15
#> 16     0       11   12   52          FALSE              TRUE       11      16
#> 17     0        9   13   88          FALSE             FALSE        9      17
#> 18     0       11   14   86          FALSE              TRUE       11      18
#> 19     0       11   20   62          FALSE              TRUE       11      19
#> 20     0       11   21   23          FALSE              TRUE       11      20
#> 21     0       11   22   47          FALSE              TRUE       11      21
#> 22     0        6   23   46          FALSE             FALSE        6      22
#> 23     0        8   25   42          FALSE             FALSE        8      23
#> 24     0        8   25   46          FALSE             FALSE        8      24
#> 25     0        8   25   98          FALSE             FALSE        8      25
#> 26     0        1   28   85          FALSE             FALSE        1      26
#> 27     0       11   30   89          FALSE              TRUE       11      27
#> 28     0       11   32   70          FALSE              TRUE       11      28
#> 29     0       11   32   77          FALSE              TRUE       11      29
#> 30     0       11   33   47          FALSE              TRUE       11      30
#> 31     0        7   35   60          FALSE             FALSE        7      31
#> 32     0        1   36   92          FALSE             FALSE        1      32
#> 33     0       11   40   54          FALSE              TRUE       11      33
#> 34     0       11   41   70          FALSE              TRUE       11      34
#> 35     0       11   43   89          FALSE              TRUE       11      35
#> 36     0       11   44   98          FALSE              TRUE       11      36
#> 37     0       11   47   97          FALSE              TRUE       11      37
#> 38     0        9   51   68          FALSE             FALSE        9      38
#> 39     0       11   55   58          FALSE              TRUE       11      39
#> 40     0        1   55   81          FALSE             FALSE        1      40
#> 41     0       11   56   66          FALSE              TRUE       11      41
#> 42     0       11   57  100          FALSE              TRUE       11      42
#> 43     0       11   60   69          FALSE              TRUE       11      43
#> 44     0        1   61   73          FALSE             FALSE        1      44
#> 45     0        9   61   95          FALSE             FALSE        9      45
#> 46     0        2   64   75          FALSE             FALSE        2      46
#> 47     0        1   64   98          FALSE             FALSE        1      47
#> 48     0       11   69   71          FALSE              TRUE       11      48
#> 49     0        8   75   90          FALSE             FALSE        8      49
#> 50     0       11   86   93          FALSE              TRUE       11      50
#> 51     0       11   91   93          FALSE              TRUE       11      51
#> 52     0        9   92   98          FALSE             FALSE        9      52
#> 53     1        5   50   91          FALSE             FALSE        4      53
#> 54     1       11   97  100          FALSE              TRUE       10      54
#> 55     2       11   38   68          FALSE              TRUE        9      55
#> 56     2       11    2   10          FALSE              TRUE        9      56
#> 57     2       11   40   69          FALSE              TRUE        9      57
#> 58     3       11   49   89          FALSE              TRUE        8      58
#> 59     4        6   11   97          FALSE             FALSE        2      59
#> 60     4       11    6   35          FALSE              TRUE        7      60
#> 61     5        8   68   87          FALSE             FALSE        3      61
#> 62     5       11   68   70          FALSE              TRUE        6      62
#> 63     5       11   49   75          FALSE              TRUE        6      63
#> 64     6       11   15   37          FALSE              TRUE        5      64
#> 65     6       11   25   91          FALSE              TRUE        5      65
#> 66     6       11   27   52          FALSE              TRUE        5      66
#> 67     6       11   77   91          FALSE              TRUE        5      67
#> 68     6       11   30   33          FALSE              TRUE        5      68
#> 69     7       11   18   37          FALSE              TRUE        4      69
#> 70     7       11    4   85          FALSE              TRUE        4      70
#> 71     8       11   45   96          FALSE              TRUE        3      71
#> 72     8       11   94   95          FALSE              TRUE        3      72
#> 73     8       11   11   80          FALSE              TRUE        3      73
#> 74     8       11   61   99          FALSE              TRUE        3      74
#> 75     8       11   38   58          FALSE              TRUE        3      75
#> 76     8       11   37   93          FALSE              TRUE        3      76
#> 77     9       11   28   31          FALSE              TRUE        2      77
#> 78     9       11   13   86          FALSE              TRUE        2      78
#> 79     9       11   43   97          FALSE              TRUE        2      79
#> 80    10       11   37   89          FALSE              TRUE        1      80

# Extract data from all simulations
as.data.frame(dx)
#>     sim onset terminus tail head onset.censored terminus.censored duration
#> 1     1     0       11    2   67          FALSE              TRUE       11
#> 2     1     0       11    3   62          FALSE              TRUE       11
#> 3     1     0       11    4   16          FALSE              TRUE       11
#> 4     1     0       11    4   58          FALSE              TRUE       11
#> 5     1     0       11    5   39          FALSE              TRUE       11
#> 6     1     0        7    6   61          FALSE             FALSE        7
#> 7     1     0        5    6   70          FALSE             FALSE        5
#> 8     1     0       11    6   79          FALSE              TRUE       11
#> 9     1     0       11    7   74          FALSE              TRUE       11
#> 10    1     0        1    8   50          FALSE             FALSE        1
#> 11    1     0        3   10   79          FALSE             FALSE        3
#> 12    1     0       11   10   81          FALSE              TRUE       11
#> 13    1     0       11   10   86          FALSE              TRUE       11
#> 14    1     0       10   11   55          FALSE             FALSE       10
#> 15    1     0       11   11   85          FALSE              TRUE       11
#> 16    1     0       11   12   52          FALSE              TRUE       11
#> 17    1     0        9   13   88          FALSE             FALSE        9
#> 18    1     0       11   14   86          FALSE              TRUE       11
#> 19    1     0       11   20   62          FALSE              TRUE       11
#> 20    1     0       11   21   23          FALSE              TRUE       11
#> 21    1     0       11   22   47          FALSE              TRUE       11
#> 22    1     0        6   23   46          FALSE             FALSE        6
#> 23    1     0        8   25   42          FALSE             FALSE        8
#> 24    1     0        8   25   46          FALSE             FALSE        8
#> 25    1     0        8   25   98          FALSE             FALSE        8
#> 26    1     0        1   28   85          FALSE             FALSE        1
#> 27    1     0       11   30   89          FALSE              TRUE       11
#> 28    1     0       11   32   70          FALSE              TRUE       11
#> 29    1     0       11   32   77          FALSE              TRUE       11
#> 30    1     0       11   33   47          FALSE              TRUE       11
#> 31    1     0        7   35   60          FALSE             FALSE        7
#> 32    1     0        1   36   92          FALSE             FALSE        1
#> 33    1     0       11   40   54          FALSE              TRUE       11
#> 34    1     0       11   41   70          FALSE              TRUE       11
#> 35    1     0       11   43   89          FALSE              TRUE       11
#> 36    1     0       11   44   98          FALSE              TRUE       11
#> 37    1     0       11   47   97          FALSE              TRUE       11
#> 38    1     0        9   51   68          FALSE             FALSE        9
#> 39    1     0       11   55   58          FALSE              TRUE       11
#> 40    1     0        1   55   81          FALSE             FALSE        1
#> 41    1     0       11   56   66          FALSE              TRUE       11
#> 42    1     0       11   57  100          FALSE              TRUE       11
#> 43    1     0       11   60   69          FALSE              TRUE       11
#> 44    1     0        1   61   73          FALSE             FALSE        1
#> 45    1     0        9   61   95          FALSE             FALSE        9
#> 46    1     0        2   64   75          FALSE             FALSE        2
#> 47    1     0        1   64   98          FALSE             FALSE        1
#> 48    1     0       11   69   71          FALSE              TRUE       11
#> 49    1     0        8   75   90          FALSE             FALSE        8
#> 50    1     0       11   86   93          FALSE              TRUE       11
#> 51    1     0       11   91   93          FALSE              TRUE       11
#> 52    1     0        9   92   98          FALSE             FALSE        9
#> 53    1     1        5   50   91          FALSE             FALSE        4
#> 54    1     1       11   97  100          FALSE              TRUE       10
#> 55    1     2       11   38   68          FALSE              TRUE        9
#> 56    1     2       11    2   10          FALSE              TRUE        9
#> 57    1     2       11   40   69          FALSE              TRUE        9
#> 58    1     3       11   49   89          FALSE              TRUE        8
#> 59    1     4        6   11   97          FALSE             FALSE        2
#> 60    1     4       11    6   35          FALSE              TRUE        7
#> 61    1     5        8   68   87          FALSE             FALSE        3
#> 62    1     5       11   68   70          FALSE              TRUE        6
#> 63    1     5       11   49   75          FALSE              TRUE        6
#> 64    1     6       11   15   37          FALSE              TRUE        5
#> 65    1     6       11   25   91          FALSE              TRUE        5
#> 66    1     6       11   27   52          FALSE              TRUE        5
#> 67    1     6       11   77   91          FALSE              TRUE        5
#> 68    1     6       11   30   33          FALSE              TRUE        5
#> 69    1     7       11   18   37          FALSE              TRUE        4
#> 70    1     7       11    4   85          FALSE              TRUE        4
#> 71    1     8       11   45   96          FALSE              TRUE        3
#> 72    1     8       11   94   95          FALSE              TRUE        3
#> 73    1     8       11   11   80          FALSE              TRUE        3
#> 74    1     8       11   61   99          FALSE              TRUE        3
#> 75    1     8       11   38   58          FALSE              TRUE        3
#> 76    1     8       11   37   93          FALSE              TRUE        3
#> 77    1     9       11   28   31          FALSE              TRUE        2
#> 78    1     9       11   13   86          FALSE              TRUE        2
#> 79    1     9       11   43   97          FALSE              TRUE        2
#> 80    1    10       11   37   89          FALSE              TRUE        1
#> 81    2     0       11    1   43          FALSE              TRUE       11
#> 82    2     0        8    2   10          FALSE             FALSE        8
#> 83    2     0        8    4   62          FALSE             FALSE        8
#> 84    2     0        3    5   83          FALSE             FALSE        3
#> 85    2     0       11    5   90          FALSE              TRUE       11
#> 86    2     0       11    6   20          FALSE              TRUE       11
#> 87    2     0       11    7   58          FALSE              TRUE       11
#> 88    2     0       11    9   69          FALSE              TRUE       11
#> 89    2     0        3    9   77          FALSE             FALSE        3
#> 90    2     0       11   11   86          FALSE              TRUE       11
#> 91    2     0       11   13   52          FALSE              TRUE       11
#> 92    2     0        9   18   90          FALSE             FALSE        9
#> 93    2     0        6   19   38          FALSE             FALSE        6
#> 94    2     0       11   19   54          FALSE              TRUE       11
#> 95    2     0        3   21   34          FALSE             FALSE        3
#> 96    2     0       11   21   54          FALSE              TRUE       11
#> 97    2     0        9   23   93          FALSE             FALSE        9
#> 98    2     0       11   24   40          FALSE              TRUE       11
#> 99    2     0       11   24   89          FALSE              TRUE       11
#> 100   2     0       11   26   36          FALSE              TRUE       11
#> 101   2     0        1   27   46          FALSE             FALSE        1
#> 102   2     0        3   28   53          FALSE             FALSE        3
#> 103   2     0       11   30   90          FALSE              TRUE       11
#> 104   2     0        5   31   76          FALSE             FALSE        5
#> 105   2     0       11   39   83          FALSE              TRUE       11
#> 106   2     0        1   46   97          FALSE             FALSE        1
#> 107   2     0       11   48   65          FALSE              TRUE       11
#> 108   2     0       11   49   68          FALSE              TRUE       11
#> 109   2     0       11   49  100          FALSE              TRUE       11
#> 110   2     0        3   51   85          FALSE             FALSE        3
#> 111   2     0       11   55   96          FALSE              TRUE       11
#> 112   2     0        5   56   70          FALSE             FALSE        5
#> 113   2     0        2   56   74          FALSE             FALSE        2
#> 114   2     0       11   57   90          FALSE              TRUE       11
#> 115   2     0       11   59   60          FALSE              TRUE       11
#> 116   2     0        4   64   87          FALSE             FALSE        4
#> 117   2     0       11   65   67          FALSE              TRUE       11
#> 118   2     0       11   65   74          FALSE              TRUE       11
#> 119   2     0       11   68   83          FALSE              TRUE       11
#> 120   2     0       11   68   84          FALSE              TRUE       11
#> 121   2     0       11   73   86          FALSE              TRUE       11
#> 122   2     0       11   74   85          FALSE              TRUE       11
#> 123   2     0       11   76   77          FALSE              TRUE       11
#> 124   2     0       11   76   81          FALSE              TRUE       11
#> 125   2     0       11   77   91          FALSE              TRUE       11
#> 126   2     0        4   78   88          FALSE             FALSE        4
#> 127   2     1        4   56   91          FALSE             FALSE        3
#> 128   2     2       11   29   55          FALSE              TRUE        9
#> 129   2     2       11   74   91          FALSE              TRUE        9
#> 130   2     2       11   17   39          FALSE              TRUE        9
#> 131   2     2        5   26   63          FALSE             FALSE        3
#> 132   2     2       11   46   54          FALSE              TRUE        9
#> 133   2     3       11   20   85          FALSE              TRUE        8
#> 134   2     3        5   60   79          FALSE             FALSE        2
#> 135   2     4       11   13   86          FALSE              TRUE        7
#> 136   2     5       11    9   37          FALSE              TRUE        6
#> 137   2     5       11   10   47          FALSE              TRUE        6
#> 138   2     5       11    7   22          FALSE              TRUE        6
#> 139   2     5       11    5   62          FALSE              TRUE        6
#> 140   2     5        9   38   97          FALSE             FALSE        4
#> 141   2     7       11   83   87          FALSE              TRUE        4
#> 142   2     7       11    7   38          FALSE              TRUE        4
#> 143   2     8       11   54   96          FALSE              TRUE        3
#> 144   2     9       11   51   73          FALSE              TRUE        2
#> 145   2     9       11   86   87          FALSE              TRUE        2
#> 146   2     9       11   33   87          FALSE              TRUE        2
#> 147   2    10       11    7   71          FALSE              TRUE        1
#> 148   3     0       11    1   40          FALSE              TRUE       11
#> 149   3     0       11    1   80          FALSE              TRUE       11
#> 150   3     0       11    2   17          FALSE              TRUE       11
#> 151   3     0        8    2   93          FALSE             FALSE        8
#> 152   3     0       11    2   97          FALSE              TRUE       11
#> 153   3     0        7    3   53          FALSE             FALSE        7
#> 154   3     0        1    3   78          FALSE             FALSE        1
#> 155   3     0       11    5   74          FALSE              TRUE       11
#> 156   3     0       11    6   56          FALSE              TRUE       11
#> 157   3     0        9    8   14          FALSE             FALSE        9
#> 158   3     0       10    8   41          FALSE             FALSE       10
#> 159   3     0       11    8   71          FALSE              TRUE       11
#> 160   3     0       11   11   16          FALSE              TRUE       11
#> 161   3     0       11   12   62          FALSE              TRUE       11
#> 162   3     0        4   13   22          FALSE             FALSE        4
#> 163   3     0        4   13   95          FALSE             FALSE        4
#> 164   3     0       11   15   21          FALSE              TRUE       11
#> 165   3     0       11   16   73          FALSE              TRUE       11
#> 166   3     0        3   17   88          FALSE             FALSE        3
#> 167   3     0       11   18   22          FALSE              TRUE       11
#> 168   3     0       11   18   23          FALSE              TRUE       11
#> 169   3     0        3   18   95          FALSE             FALSE        3
#> 170   3     0       11   19   28          FALSE              TRUE       11
#> 171   3     0       11   22   68          FALSE              TRUE       11
#> 172   3     0       11   24   67          FALSE              TRUE       11
#> 173   3     0        6   24   85          FALSE             FALSE        6
#> 174   3     0       11   25   82          FALSE              TRUE       11
#> 175   3     0       11   26   27          FALSE              TRUE       11
#> 176   3     0        7   26  100          FALSE             FALSE        7
#> 177   3     0       11   29   84          FALSE              TRUE       11
#> 178   3     0       11   32   46          FALSE              TRUE       11
#> 179   3     0        6   32   87          FALSE             FALSE        6
#> 180   3     0       11   33   62          FALSE              TRUE       11
#> 181   3     0        3   33   79          FALSE             FALSE        3
#> 182   3     0       11   35   64          FALSE              TRUE       11
#> 183   3     0        4   36   48          FALSE             FALSE        4
#> 184   3     0       11   37   73          FALSE              TRUE       11
#> 185   3     0        8   38   60          FALSE             FALSE        8
#> 186   3     0       11   39   73          FALSE              TRUE       11
#> 187   3     0       11   41   76          FALSE              TRUE       11
#> 188   3     0       11   41   78          FALSE              TRUE       11
#> 189   3     0       11   42   62          FALSE              TRUE       11
#> 190   3     0       11   42   66          FALSE              TRUE       11
#> 191   3     0        6   45   53          FALSE             FALSE        6
#> 192   3     0       11   46   59          FALSE              TRUE       11
#> 193   3     0       11   50   95          FALSE              TRUE       11
#> 194   3     0        9   55   85          FALSE             FALSE        9
#> 195   3     0       11   56   70          FALSE              TRUE       11
#> 196   3     0        4   60   65          FALSE             FALSE        4
#> 197   3     0        6   61   65          FALSE             FALSE        6
#> 198   3     0        3   65   81          FALSE             FALSE        3
#> 199   3     0        6   70   76          FALSE             FALSE        6
#> 200   3     0        3   77   80          FALSE             FALSE        3
#> 201   3     0        6   77   96          FALSE             FALSE        6
#> 202   3     0       11   80   84          FALSE              TRUE       11
#> 203   3     0       10   86   93          FALSE             FALSE       10
#> 204   3     1       11   27   82          FALSE              TRUE       10
#> 205   3     1       11   53   56          FALSE              TRUE       10
#> 206   3     1       11   24   89          FALSE              TRUE       10
#> 207   3     1       11    3   28          FALSE              TRUE       10
#> 208   3     2       11   21   38          FALSE              TRUE        9
#> 209   3     2        6   23   95          FALSE             FALSE        4
#> 210   3     2       11   33   44          FALSE              TRUE        9
#> 211   3     2       10   66   98          FALSE             FALSE        8
#> 212   3     3        6   70   93          FALSE             FALSE        3
#> 213   3     3       11   27   36          FALSE              TRUE        8
#> 214   3     4       11    2   57          FALSE              TRUE        7
#> 215   3     4       11    3   96          FALSE              TRUE        7
#> 216   3     4       11   58   98          FALSE              TRUE        7
#> 217   3     5        9    4   66          FALSE             FALSE        4
#> 218   3     5       11   32   77          FALSE              TRUE        6
#> 219   3     5       11   13   19          FALSE              TRUE        6
#> 220   3     5        8   58   92          FALSE             FALSE        3
#> 221   3     6        7   11   69          FALSE             FALSE        1
#> 222   3     6       11   29   43          FALSE              TRUE        5
#> 223   3     6       11   13   58          FALSE              TRUE        5
#> 224   3     7       11   26   99          FALSE              TRUE        4
#> 225   3     8       11   90   95          FALSE              TRUE        3
#> 226   3     8       11   43   68          FALSE              TRUE        3
#> 227   3     8       11   50   58          FALSE              TRUE        3
#> 228   3     9       11    4   54          FALSE              TRUE        2
#> 229   3     9       10   30   31          FALSE             FALSE        1
#> 230   3     9       11    9   61          FALSE              TRUE        2
#> 231   3    10       11   89   94          FALSE              TRUE        1
#> 232   3    10       11   51   58          FALSE              TRUE        1
#>     edge.id
#> 1         1
#> 2         2
#> 3         3
#> 4         4
#> 5         5
#> 6         6
#> 7         7
#> 8         8
#> 9         9
#> 10       10
#> 11       11
#> 12       12
#> 13       13
#> 14       14
#> 15       15
#> 16       16
#> 17       17
#> 18       18
#> 19       19
#> 20       20
#> 21       21
#> 22       22
#> 23       23
#> 24       24
#> 25       25
#> 26       26
#> 27       27
#> 28       28
#> 29       29
#> 30       30
#> 31       31
#> 32       32
#> 33       33
#> 34       34
#> 35       35
#> 36       36
#> 37       37
#> 38       38
#> 39       39
#> 40       40
#> 41       41
#> 42       42
#> 43       43
#> 44       44
#> 45       45
#> 46       46
#> 47       47
#> 48       48
#> 49       49
#> 50       50
#> 51       51
#> 52       52
#> 53       53
#> 54       54
#> 55       55
#> 56       56
#> 57       57
#> 58       58
#> 59       59
#> 60       60
#> 61       61
#> 62       62
#> 63       63
#> 64       64
#> 65       65
#> 66       66
#> 67       67
#> 68       68
#> 69       69
#> 70       70
#> 71       71
#> 72       72
#> 73       73
#> 74       74
#> 75       75
#> 76       76
#> 77       77
#> 78       78
#> 79       79
#> 80       80
#> 81        1
#> 82        2
#> 83        3
#> 84        4
#> 85        5
#> 86        6
#> 87        7
#> 88        8
#> 89        9
#> 90       10
#> 91       11
#> 92       12
#> 93       13
#> 94       14
#> 95       15
#> 96       16
#> 97       17
#> 98       18
#> 99       19
#> 100      20
#> 101      21
#> 102      22
#> 103      23
#> 104      24
#> 105      25
#> 106      26
#> 107      27
#> 108      28
#> 109      29
#> 110      30
#> 111      31
#> 112      32
#> 113      33
#> 114      34
#> 115      35
#> 116      36
#> 117      37
#> 118      38
#> 119      39
#> 120      40
#> 121      41
#> 122      42
#> 123      43
#> 124      44
#> 125      45
#> 126      46
#> 127      47
#> 128      48
#> 129      49
#> 130      50
#> 131      51
#> 132      52
#> 133      53
#> 134      54
#> 135      55
#> 136      56
#> 137      57
#> 138      58
#> 139      59
#> 140      60
#> 141      61
#> 142      62
#> 143      63
#> 144      64
#> 145      65
#> 146      66
#> 147      67
#> 148       1
#> 149       2
#> 150       3
#> 151       4
#> 152       5
#> 153       6
#> 154       7
#> 155       8
#> 156       9
#> 157      10
#> 158      11
#> 159      12
#> 160      13
#> 161      14
#> 162      15
#> 163      16
#> 164      17
#> 165      18
#> 166      19
#> 167      20
#> 168      21
#> 169      22
#> 170      23
#> 171      24
#> 172      25
#> 173      26
#> 174      27
#> 175      28
#> 176      29
#> 177      30
#> 178      31
#> 179      32
#> 180      33
#> 181      34
#> 182      35
#> 183      36
#> 184      37
#> 185      38
#> 186      39
#> 187      40
#> 188      41
#> 189      42
#> 190      43
#> 191      44
#> 192      45
#> 193      46
#> 194      47
#> 195      48
#> 196      49
#> 197      50
#> 198      51
#> 199      52
#> 200      53
#> 201      54
#> 202      55
#> 203      56
#> 204      57
#> 205      58
#> 206      59
#> 207      60
#> 208      61
#> 209      62
#> 210      63
#> 211      64
#> 212      65
#> 213      66
#> 214      67
#> 215      68
#> 216      69
#> 217      70
#> 218      71
#> 219      72
#> 220      73
#> 221      74
#> 222      75
#> 223      76
#> 224      77
#> 225      78
#> 226      79
#> 227      80
#> 228      81
#> 229      82
#> 230      83
#> 231      84
#> 232      85