Extract Timed Edgelists for netdx Objects

This function extracts timed edgelists for objects of class netdx into a data frame using the generic as.data.frame function.

Usage

# S3 method for netdx
as.data.frame(x, row.names = NULL, optional = FALSE, sim, ...)

Arguments

x

An EpiModel object of class netdx.

row.names

See as.data.frame.default.

optional

See as.data.frame.default.

sim

The simulation number to output. If not specified, then data from all simulations will be output.

...

See as.data.frame.default.

Value

A data frame containing the data from x.

Examples

# Initialize and parameterize the network model
nw <- network_initialize(n = 100)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)

# Model estimation
est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)
#> Starting maximum pseudolikelihood estimation (MPLE):
#> Obtaining the responsible dyads.
#> Evaluating the predictor and response matrix.
#> Maximizing the pseudolikelihood.
#> Finished MPLE.

# Simulate the network with netdx
dx <- netdx(est, nsims = 3, nsteps = 10, keep.tedgelist = TRUE,
            verbose = FALSE)

# Extract data from the first simulation
as.data.frame(dx, sim = 1)
#>    onset terminus tail head onset.censored terminus.censored duration edge.id
#> 1      0        6    1   26          FALSE             FALSE        6       1
#> 2      0       11    1   83          FALSE              TRUE       11       2
#> 3      0        1    4    6          FALSE             FALSE        1       3
#> 4      0       11    4   75          FALSE              TRUE       11       4
#> 5      0       11    8   76          FALSE              TRUE       11       5
#> 6      0       11   11   76          FALSE              TRUE       11       6
#> 7      0        6   12   78          FALSE             FALSE        6       7
#> 8      0       10   12  100          FALSE             FALSE       10       8
#> 9      0       11   13   21          FALSE              TRUE       11       9
#> 10     0       11   13   81          FALSE              TRUE       11      10
#> 11     0        8   15   61          FALSE             FALSE        8      11
#> 12     0       11   18   87          FALSE              TRUE       11      12
#> 13     0       11   19   76          FALSE              TRUE       11      13
#> 14     0        6   19   90          FALSE             FALSE        6      14
#> 15     0        4   21   30          FALSE             FALSE        4      15
#> 16     0       11   21   89          FALSE              TRUE       11      16
#> 17     0        7   21   94          FALSE             FALSE        7      17
#> 18     0        4   22   26          FALSE             FALSE        4      18
#> 19     0       11   23   97          FALSE              TRUE       11      19
#> 20     0       11   24   38          FALSE              TRUE       11      20
#> 21     0       11   25   63          FALSE              TRUE       11      21
#> 22     0       11   26   29          FALSE              TRUE       11      22
#> 23     0       11   27   33          FALSE              TRUE       11      23
#> 24     0        6   28   41          FALSE             FALSE        6      24
#> 25     0       11   36   92          FALSE              TRUE       11      25
#> 26     0        7   38   64          FALSE             FALSE        7      26
#> 27     0       11   40   81          FALSE              TRUE       11      27
#> 28     0       10   41   58          FALSE             FALSE       10      28
#> 29     0        7   41   77          FALSE             FALSE        7      29
#> 30     0       11   45   88          FALSE              TRUE       11      30
#> 31     0        1   49   71          FALSE             FALSE        1      31
#> 32     0       11   49   93          FALSE              TRUE       11      32
#> 33     0       11   53   89          FALSE              TRUE       11      33
#> 34     0       11   54   91          FALSE              TRUE       11      34
#> 35     0       11   55   76          FALSE              TRUE       11      35
#> 36     0        5   56   64          FALSE             FALSE        5      36
#> 37     0       11   62   98          FALSE              TRUE       11      37
#> 38     0        6   63   86          FALSE             FALSE        6      38
#> 39     0       11   67   77          FALSE              TRUE       11      39
#> 40     0        5   67   89          FALSE             FALSE        5      40
#> 41     0       11   70   95          FALSE              TRUE       11      41
#> 42     0       11   77   96          FALSE              TRUE       11      42
#> 43     0       11   80   90          FALSE              TRUE       11      43
#> 44     0       11   81   98          FALSE              TRUE       11      44
#> 45     1       11   30   53          FALSE              TRUE       10      45
#> 46     1       11   31   49          FALSE              TRUE       10      46
#> 47     2        3   59   60          FALSE             FALSE        1      47
#> 48     2       11   13   17          FALSE              TRUE        9      48
#> 49     2        4   22   77          FALSE             FALSE        2      49
#> 50     2       11   81   89          FALSE              TRUE        9      50
#> 51     2        4   24   69          FALSE             FALSE        2      51
#> 52     3        5   40   73          FALSE             FALSE        2      52
#> 53     4        5   57   76          FALSE             FALSE        1      53
#> 54     4       11   21   47          FALSE              TRUE        7      54
#> 55     4       11    9   25          FALSE              TRUE        7      55
#> 56     4       11   65   95          FALSE              TRUE        7      56
#> 57     5       11   57   93          FALSE              TRUE        6      57
#> 58     5        6   27   90          FALSE             FALSE        1      58
#> 59     6       11   44   85          FALSE              TRUE        5      59
#> 60     6       11   64   72          FALSE              TRUE        5      60
#> 61     7       11   23   92          FALSE              TRUE        4      61
#> 62     7       11    5   80          FALSE              TRUE        4      62
#> 63     8       11    6   71          FALSE              TRUE        3      63
#> 64     8       11   47   82          FALSE              TRUE        3      64
#> 65     8        9   46   68          FALSE             FALSE        1      65
#> 66     8       11   54   84          FALSE              TRUE        3      66
#> 67     9       11   22   52          FALSE              TRUE        2      67
#> 68    10       11   25   58          FALSE              TRUE        1      68
#> 69    10       11   41   45          FALSE              TRUE        1      69
#> 70    10       11   14   46          FALSE              TRUE        1      70
#> 71    10       11   83   92          FALSE              TRUE        1      71

# Extract data from all simulations
as.data.frame(dx)
#>     sim onset terminus tail head onset.censored terminus.censored duration
#> 1     1     0        6    1   26          FALSE             FALSE        6
#> 2     1     0       11    1   83          FALSE              TRUE       11
#> 3     1     0        1    4    6          FALSE             FALSE        1
#> 4     1     0       11    4   75          FALSE              TRUE       11
#> 5     1     0       11    8   76          FALSE              TRUE       11
#> 6     1     0       11   11   76          FALSE              TRUE       11
#> 7     1     0        6   12   78          FALSE             FALSE        6
#> 8     1     0       10   12  100          FALSE             FALSE       10
#> 9     1     0       11   13   21          FALSE              TRUE       11
#> 10    1     0       11   13   81          FALSE              TRUE       11
#> 11    1     0        8   15   61          FALSE             FALSE        8
#> 12    1     0       11   18   87          FALSE              TRUE       11
#> 13    1     0       11   19   76          FALSE              TRUE       11
#> 14    1     0        6   19   90          FALSE             FALSE        6
#> 15    1     0        4   21   30          FALSE             FALSE        4
#> 16    1     0       11   21   89          FALSE              TRUE       11
#> 17    1     0        7   21   94          FALSE             FALSE        7
#> 18    1     0        4   22   26          FALSE             FALSE        4
#> 19    1     0       11   23   97          FALSE              TRUE       11
#> 20    1     0       11   24   38          FALSE              TRUE       11
#> 21    1     0       11   25   63          FALSE              TRUE       11
#> 22    1     0       11   26   29          FALSE              TRUE       11
#> 23    1     0       11   27   33          FALSE              TRUE       11
#> 24    1     0        6   28   41          FALSE             FALSE        6
#> 25    1     0       11   36   92          FALSE              TRUE       11
#> 26    1     0        7   38   64          FALSE             FALSE        7
#> 27    1     0       11   40   81          FALSE              TRUE       11
#> 28    1     0       10   41   58          FALSE             FALSE       10
#> 29    1     0        7   41   77          FALSE             FALSE        7
#> 30    1     0       11   45   88          FALSE              TRUE       11
#> 31    1     0        1   49   71          FALSE             FALSE        1
#> 32    1     0       11   49   93          FALSE              TRUE       11
#> 33    1     0       11   53   89          FALSE              TRUE       11
#> 34    1     0       11   54   91          FALSE              TRUE       11
#> 35    1     0       11   55   76          FALSE              TRUE       11
#> 36    1     0        5   56   64          FALSE             FALSE        5
#> 37    1     0       11   62   98          FALSE              TRUE       11
#> 38    1     0        6   63   86          FALSE             FALSE        6
#> 39    1     0       11   67   77          FALSE              TRUE       11
#> 40    1     0        5   67   89          FALSE             FALSE        5
#> 41    1     0       11   70   95          FALSE              TRUE       11
#> 42    1     0       11   77   96          FALSE              TRUE       11
#> 43    1     0       11   80   90          FALSE              TRUE       11
#> 44    1     0       11   81   98          FALSE              TRUE       11
#> 45    1     1       11   30   53          FALSE              TRUE       10
#> 46    1     1       11   31   49          FALSE              TRUE       10
#> 47    1     2        3   59   60          FALSE             FALSE        1
#> 48    1     2       11   13   17          FALSE              TRUE        9
#> 49    1     2        4   22   77          FALSE             FALSE        2
#> 50    1     2       11   81   89          FALSE              TRUE        9
#> 51    1     2        4   24   69          FALSE             FALSE        2
#> 52    1     3        5   40   73          FALSE             FALSE        2
#> 53    1     4        5   57   76          FALSE             FALSE        1
#> 54    1     4       11   21   47          FALSE              TRUE        7
#> 55    1     4       11    9   25          FALSE              TRUE        7
#> 56    1     4       11   65   95          FALSE              TRUE        7
#> 57    1     5       11   57   93          FALSE              TRUE        6
#> 58    1     5        6   27   90          FALSE             FALSE        1
#> 59    1     6       11   44   85          FALSE              TRUE        5
#> 60    1     6       11   64   72          FALSE              TRUE        5
#> 61    1     7       11   23   92          FALSE              TRUE        4
#> 62    1     7       11    5   80          FALSE              TRUE        4
#> 63    1     8       11    6   71          FALSE              TRUE        3
#> 64    1     8       11   47   82          FALSE              TRUE        3
#> 65    1     8        9   46   68          FALSE             FALSE        1
#> 66    1     8       11   54   84          FALSE              TRUE        3
#> 67    1     9       11   22   52          FALSE              TRUE        2
#> 68    1    10       11   25   58          FALSE              TRUE        1
#> 69    1    10       11   41   45          FALSE              TRUE        1
#> 70    1    10       11   14   46          FALSE              TRUE        1
#> 71    1    10       11   83   92          FALSE              TRUE        1
#> 72    2     0       11    1   86          FALSE              TRUE       11
#> 73    2     0       10    3   69          FALSE             FALSE       10
#> 74    2     0       11    3   81          FALSE              TRUE       11
#> 75    2     0       11    4   87          FALSE              TRUE       11
#> 76    2     0       11    5   14          FALSE              TRUE       11
#> 77    2     0        4    6   78          FALSE             FALSE        4
#> 78    2     0        4    7   26          FALSE             FALSE        4
#> 79    2     0        4    8   38          FALSE             FALSE        4
#> 80    2     0       11    8   73          FALSE              TRUE       11
#> 81    2     0       11    8   85          FALSE              TRUE       11
#> 82    2     0        1    9   24          FALSE             FALSE        1
#> 83    2     0       11    9   72          FALSE              TRUE       11
#> 84    2     0       11    9   94          FALSE              TRUE       11
#> 85    2     0        1   10   33          FALSE             FALSE        1
#> 86    2     0       11   10   61          FALSE              TRUE       11
#> 87    2     0       11   11   61          FALSE              TRUE       11
#> 88    2     0       11   14   72          FALSE              TRUE       11
#> 89    2     0       11   15   37          FALSE              TRUE       11
#> 90    2     0       11   15   90          FALSE              TRUE       11
#> 91    2     0       11   16   69          FALSE              TRUE       11
#> 92    2     0       11   20   71          FALSE              TRUE       11
#> 93    2     0       11   22   33          FALSE              TRUE       11
#> 94    2     0        1   25   56          FALSE             FALSE        1
#> 95    2     0        5   27   58          FALSE             FALSE        5
#> 96    2     0       11   27   90          FALSE              TRUE       11
#> 97    2     0        3   31   73          FALSE             FALSE        3
#> 98    2     0       11   34   38          FALSE              TRUE       11
#> 99    2     0        9   35   74          FALSE             FALSE        9
#> 100   2     0        5   36   79          FALSE             FALSE        5
#> 101   2     0       11   37   40          FALSE              TRUE       11
#> 102   2     0       11   38   47          FALSE              TRUE       11
#> 103   2     0        6   39   63          FALSE             FALSE        6
#> 104   2     0       11   39   92          FALSE              TRUE       11
#> 105   2     0        5   39   98          FALSE             FALSE        5
#> 106   2     0        2   41   80          FALSE             FALSE        2
#> 107   2     0        5   41   83          FALSE             FALSE        5
#> 108   2     0        1   42   95          FALSE             FALSE        1
#> 109   2     0       11   42   99          FALSE              TRUE       11
#> 110   2     0       11   44   83          FALSE              TRUE       11
#> 111   2     0       11   45   63          FALSE              TRUE       11
#> 112   2     0       11   50   61          FALSE              TRUE       11
#> 113   2     0        8   54   67          FALSE             FALSE        8
#> 114   2     0        6   54   79          FALSE             FALSE        6
#> 115   2     0        5   54   86          FALSE             FALSE        5
#> 116   2     0        1   55   59          FALSE             FALSE        1
#> 117   2     0       11   55   60          FALSE              TRUE       11
#> 118   2     0        1   55   95          FALSE             FALSE        1
#> 119   2     0       11   56   97          FALSE              TRUE       11
#> 120   2     0       10   58   79          FALSE             FALSE       10
#> 121   2     0       11   58   91          FALSE              TRUE       11
#> 122   2     0       11   61   66          FALSE              TRUE       11
#> 123   2     0       11   61   78          FALSE              TRUE       11
#> 124   2     0       11   62   83          FALSE              TRUE       11
#> 125   2     0        3   69   70          FALSE             FALSE        3
#> 126   2     0       11   73   81          FALSE              TRUE       11
#> 127   2     0       11   77   80          FALSE              TRUE       11
#> 128   2     0       11   81   82          FALSE              TRUE       11
#> 129   2     0       11   81   95          FALSE              TRUE       11
#> 130   2     1        8   22   57          FALSE             FALSE        7
#> 131   2     1       11   23   36          FALSE              TRUE       10
#> 132   2     1        3    5   34          FALSE             FALSE        2
#> 133   2     2       11   26   85          FALSE              TRUE        9
#> 134   2     2       11   60   82          FALSE              TRUE        9
#> 135   2     2       11    3   84          FALSE              TRUE        9
#> 136   2     3       11   35   73          FALSE              TRUE        8
#> 137   2     4       11   66   71          FALSE              TRUE        7
#> 138   2     4       11   55   98          FALSE              TRUE        7
#> 139   2     5       11    1   32          FALSE              TRUE        6
#> 140   2     5       11   39   91          FALSE              TRUE        6
#> 141   2     5       11    5   64          FALSE              TRUE        6
#> 142   2     7       11   54   88          FALSE              TRUE        4
#> 143   2     7        9   60   81          FALSE             FALSE        2
#> 144   2     7       11   41   90          FALSE              TRUE        4
#> 145   2     8       11   12   94          FALSE              TRUE        3
#> 146   2     8       11   23   91          FALSE              TRUE        3
#> 147   2     8       11   40   66          FALSE              TRUE        3
#> 148   2     8       11    8   47          FALSE              TRUE        3
#> 149   2     8       11   92  100          FALSE              TRUE        3
#> 150   2     9       11   89   92          FALSE              TRUE        2
#> 151   2     9       10   30   89          FALSE             FALSE        1
#> 152   2    10       11   45  100          FALSE              TRUE        1
#> 153   3     0       11    1   34          FALSE              TRUE       11
#> 154   3     0       11    3   42          FALSE              TRUE       11
#> 155   3     0       11    4   96          FALSE              TRUE       11
#> 156   3     0        5    5    6          FALSE             FALSE        5
#> 157   3     0       11    5   99          FALSE              TRUE       11
#> 158   3     0       11    7   35          FALSE              TRUE       11
#> 159   3     0       11    7   73          FALSE              TRUE       11
#> 160   3     0       11    7   99          FALSE              TRUE       11
#> 161   3     0       11   10   67          FALSE              TRUE       11
#> 162   3     0        1   12   39          FALSE             FALSE        1
#> 163   3     0       11   12   85          FALSE              TRUE       11
#> 164   3     0       11   13   50          FALSE              TRUE       11
#> 165   3     0       11   13   87          FALSE              TRUE       11
#> 166   3     0        2   15   30          FALSE             FALSE        2
#> 167   3     0       11   15   70          FALSE              TRUE       11
#> 168   3     0       11   16   60          FALSE              TRUE       11
#> 169   3     0       11   17   25          FALSE              TRUE       11
#> 170   3     0       11   17   42          FALSE              TRUE       11
#> 171   3     0       11   17   88          FALSE              TRUE       11
#> 172   3     0       10   18   99          FALSE             FALSE       10
#> 173   3     0       11   19   32          FALSE              TRUE       11
#> 174   3     0       11   20   46          FALSE              TRUE       11
#> 175   3     0        3   22   75          FALSE             FALSE        3
#> 176   3     0        9   24   47          FALSE             FALSE        9
#> 177   3     0       11   24   49          FALSE              TRUE       11
#> 178   3     0        7   25   71          FALSE             FALSE        7
#> 179   3     0       11   27   98          FALSE              TRUE       11
#> 180   3     0        5   28   29          FALSE             FALSE        5
#> 181   3     0        5   29   53          FALSE             FALSE        5
#> 182   3     0        1   30   49          FALSE             FALSE        1
#> 183   3     0        1   33   81          FALSE             FALSE        1
#> 184   3     0       11   35   45          FALSE              TRUE       11
#> 185   3     0        7   36   51          FALSE             FALSE        7
#> 186   3     0        5   36   69          FALSE             FALSE        5
#> 187   3     0        8   38   81          FALSE             FALSE        8
#> 188   3     0       10   40   97          FALSE             FALSE       10
#> 189   3     0        1   41   62          FALSE             FALSE        1
#> 190   3     0       11   43   54          FALSE              TRUE       11
#> 191   3     0       11   43   85          FALSE              TRUE       11
#> 192   3     0        1   43   95          FALSE             FALSE        1
#> 193   3     0        5   44   49          FALSE             FALSE        5
#> 194   3     0       11   46   53          FALSE              TRUE       11
#> 195   3     0       11   48   95          FALSE              TRUE       11
#> 196   3     0        6   48   97          FALSE             FALSE        6
#> 197   3     0        5   50   65          FALSE             FALSE        5
#> 198   3     0       11   53   54          FALSE              TRUE       11
#> 199   3     0        2   54   88          FALSE             FALSE        2
#> 200   3     0       11   55   90          FALSE              TRUE       11
#> 201   3     0       11   56   86          FALSE              TRUE       11
#> 202   3     0       11   56   98          FALSE              TRUE       11
#> 203   3     0       11   59   95          FALSE              TRUE       11
#> 204   3     0        8   60   64          FALSE             FALSE        8
#> 205   3     0       11   63   78          FALSE              TRUE       11
#> 206   3     0       11   63   88          FALSE              TRUE       11
#> 207   3     0       11   79   94          FALSE              TRUE       11
#> 208   3     0        4   79   99          FALSE             FALSE        4
#> 209   3     0       11   84   92          FALSE              TRUE       11
#> 210   3     0        7   89   93          FALSE             FALSE        7
#> 211   3     0        1   90   94          FALSE             FALSE        1
#> 212   3     0       11   90  100          FALSE              TRUE       11
#> 213   3     0       11   93   98          FALSE              TRUE       11
#> 214   3     1        3   48   56          FALSE             FALSE        2
#> 215   3     1       11   51   68          FALSE              TRUE       10
#> 216   3     2        6   10   14          FALSE             FALSE        4
#> 217   3     2       11   24   82          FALSE              TRUE        9
#> 218   3     2       11    3   46          FALSE              TRUE        9
#> 219   3     2       11   12   75          FALSE              TRUE        9
#> 220   3     2       10   31   56          FALSE             FALSE        8
#> 221   3     2       11   45   70          FALSE              TRUE        9
#> 222   3     3       11   10   60          FALSE              TRUE        8
#> 223   3     3       11   41   73          FALSE              TRUE        8
#> 224   3     4       11   26   54          FALSE              TRUE        7
#> 225   3     4       11    6   46          FALSE              TRUE        7
#> 226   3     4        7   35   94          FALSE             FALSE        3
#> 227   3     4       11   85   86          FALSE              TRUE        7
#> 228   3     6       11   15   91          FALSE              TRUE        5
#> 229   3     6        8    8   27          FALSE             FALSE        2
#> 230   3     7        9   21   28          FALSE             FALSE        2
#> 231   3     7       11    1   61          FALSE              TRUE        4
#> 232   3     7        9   17   79          FALSE             FALSE        2
#> 233   3     8       11   85   92          FALSE              TRUE        3
#> 234   3     8       11   31   63          FALSE              TRUE        3
#> 235   3     8       11   32   61          FALSE              TRUE        3
#> 236   3     9       11   37   99          FALSE              TRUE        2
#> 237   3     9       11   73   85          FALSE              TRUE        2
#> 238   3     9       11   11   58          FALSE              TRUE        2
#>     edge.id
#> 1         1
#> 2         2
#> 3         3
#> 4         4
#> 5         5
#> 6         6
#> 7         7
#> 8         8
#> 9         9
#> 10       10
#> 11       11
#> 12       12
#> 13       13
#> 14       14
#> 15       15
#> 16       16
#> 17       17
#> 18       18
#> 19       19
#> 20       20
#> 21       21
#> 22       22
#> 23       23
#> 24       24
#> 25       25
#> 26       26
#> 27       27
#> 28       28
#> 29       29
#> 30       30
#> 31       31
#> 32       32
#> 33       33
#> 34       34
#> 35       35
#> 36       36
#> 37       37
#> 38       38
#> 39       39
#> 40       40
#> 41       41
#> 42       42
#> 43       43
#> 44       44
#> 45       45
#> 46       46
#> 47       47
#> 48       48
#> 49       49
#> 50       50
#> 51       51
#> 52       52
#> 53       53
#> 54       54
#> 55       55
#> 56       56
#> 57       57
#> 58       58
#> 59       59
#> 60       60
#> 61       61
#> 62       62
#> 63       63
#> 64       64
#> 65       65
#> 66       66
#> 67       67
#> 68       68
#> 69       69
#> 70       70
#> 71       71
#> 72        1
#> 73        2
#> 74        3
#> 75        4
#> 76        5
#> 77        6
#> 78        7
#> 79        8
#> 80        9
#> 81       10
#> 82       11
#> 83       12
#> 84       13
#> 85       14
#> 86       15
#> 87       16
#> 88       17
#> 89       18
#> 90       19
#> 91       20
#> 92       21
#> 93       22
#> 94       23
#> 95       24
#> 96       25
#> 97       26
#> 98       27
#> 99       28
#> 100      29
#> 101      30
#> 102      31
#> 103      32
#> 104      33
#> 105      34
#> 106      35
#> 107      36
#> 108      37
#> 109      38
#> 110      39
#> 111      40
#> 112      41
#> 113      42
#> 114      43
#> 115      44
#> 116      45
#> 117      46
#> 118      47
#> 119      48
#> 120      49
#> 121      50
#> 122      51
#> 123      52
#> 124      53
#> 125      54
#> 126      55
#> 127      56
#> 128      57
#> 129      58
#> 130      59
#> 131      60
#> 132      61
#> 133      62
#> 134      63
#> 135      64
#> 136      65
#> 137      66
#> 138      67
#> 139      68
#> 140      69
#> 141      70
#> 142      71
#> 143      72
#> 144      73
#> 145      74
#> 146      75
#> 147      76
#> 148      77
#> 149      78
#> 150      79
#> 151      80
#> 152      81
#> 153       1
#> 154       2
#> 155       3
#> 156       4
#> 157       5
#> 158       6
#> 159       7
#> 160       8
#> 161       9
#> 162      10
#> 163      11
#> 164      12
#> 165      13
#> 166      14
#> 167      15
#> 168      16
#> 169      17
#> 170      18
#> 171      19
#> 172      20
#> 173      21
#> 174      22
#> 175      23
#> 176      24
#> 177      25
#> 178      26
#> 179      27
#> 180      28
#> 181      29
#> 182      30
#> 183      31
#> 184      32
#> 185      33
#> 186      34
#> 187      35
#> 188      36
#> 189      37
#> 190      38
#> 191      39
#> 192      40
#> 193      41
#> 194      42
#> 195      43
#> 196      44
#> 197      45
#> 198      46
#> 199      47
#> 200      48
#> 201      49
#> 202      50
#> 203      51
#> 204      52
#> 205      53
#> 206      54
#> 207      55
#> 208      56
#> 209      57
#> 210      58
#> 211      59
#> 212      60
#> 213      61
#> 214      62
#> 215      63
#> 216      64
#> 217      65
#> 218      66
#> 219      67
#> 220      68
#> 221      69
#> 222      70
#> 223      71
#> 224      72
#> 225      73
#> 226      74
#> 227      75
#> 228      76
#> 229      77
#> 230      78
#> 231      79
#> 232      80
#> 233      81
#> 234      82
#> 235      83
#> 236      84
#> 237      85
#> 238      86