Extract Network Statistics from netsim or netdx Object

Extracts network statistics from a network epidemic model simulated with netsim or a network diagnostics object simulated with netdx. Statistics can be returned either as a single data frame or as a list of matrices (one matrix for each simulation).

Usage

get_nwstats(x, sim, network = 1, mode = c("data.frame", "list"))

Arguments

x

An EpiModel object of class netsim or netdx.

sim

A vector of simulation numbers from the extracted object.

network

Network number, for netsim objects with multiple overlapping networks (advanced use, and not applicable to netdx objects).

mode

Either "data.frame" or "list", indicating the desired output.

Value

A data frame or list of matrices containing the network statistics.

Examples

# Two-group Bernoulli random graph TERGM
nw <- network_initialize(n = 100)
nw <- set_vertex_attribute(nw, "group", rep(1:2, each = 50))
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)
est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)
#> Starting maximum pseudolikelihood estimation (MPLE):
#> Obtaining the responsible dyads.
#> Evaluating the predictor and response matrix.
#> Maximizing the pseudolikelihood.
#> Finished MPLE.

dx <- netdx(est, nsim = 3, nsteps = 10, verbose = FALSE,
            nwstats.formula = ~edges + isolates)
get_nwstats(dx)
#>    time sim edges isolates
#> 1     1   1    40       48
#> 2     2   1    42       44
#> 3     3   1    41       45
#> 4     4   1    43       43
#> 5     5   1    44       42
#> 6     6   1    45       41
#> 7     7   1    46       38
#> 8     8   1    45       38
#> 9     9   1    47       36
#> 10   10   1    52       36
#> 11    1   2    49       40
#> 12    2   2    48       42
#> 13    3   2    50       39
#> 14    4   2    47       41
#> 15    5   2    49       37
#> 16    6   2    52       39
#> 17    7   2    53       35
#> 18    8   2    57       31
#> 19    9   2    58       33
#> 20   10   2    57       33
#> 21    1   3    45       42
#> 22    2   3    40       50
#> 23    3   3    40       51
#> 24    4   3    42       48
#> 25    5   3    41       50
#> 26    6   3    41       50
#> 27    7   3    42       50
#> 28    8   3    44       46
#> 29    9   3    43       47
#> 30   10   3    48       39
get_nwstats(dx, sim = 1)
#>    time sim edges isolates
#> 1     1   1    40       48
#> 2     2   1    42       44
#> 3     3   1    41       45
#> 4     4   1    43       43
#> 5     5   1    44       42
#> 6     6   1    45       41
#> 7     7   1    46       38
#> 8     8   1    45       38
#> 9     9   1    47       36
#> 10   10   1    52       36

# SI epidemic model
param <- param.net(inf.prob = 0.3, inf.prob.g2 = 0.15)
init <- init.net(i.num = 10, i.num.g2 = 10)
control <- control.net(type = "SI", nsteps = 10, nsims = 3,
                       nwstats.formula = ~edges + meandeg + degree(0:5),
                       verbose = FALSE)
mod <- netsim(est, param, init, control)

# Extract the network statistics from all or sets of simulations
get_nwstats(mod)
#>    time sim edges meandeg degree0 degree1 degree2 degree3 degree4 degree5
#> 1     1   1    53    1.06      33      40      17       8       2       0
#> 2     2   1    52    1.04      35      38      17       8       2       0
#> 3     3   1    52    1.04      35      38      17       8       2       0
#> 4     4   1    53    1.06      36      37      15       9       3       0
#> 5     5   1    51    1.02      38      37      13       9       3       0
#> 6     6   1    52    1.04      37      36      16       8       3       0
#> 7     7   1    49    0.98      41      36      11       8       4       0
#> 8     8   1    50    1.00      39      37      14       6       3       1
#> 9     9   1    48    0.96      42      33      16       6       2       1
#> 10   10   1    47    0.94      44      30      18       5       2       1
#> 11    1   2    55    1.10      44      23      18      10       4       1
#> 12    2   2    54    1.08      44      26      15      10       4       0
#> 13    3   2    55    1.10      44      26      14      10       5       0
#> 14    4   2    50    1.00      45      30      11       9       4       1
#> 15    5   2    49    0.98      43      34      10       9       3       1
#> 16    6   2    51    1.02      40      36      10      11       2       1
#> 17    7   2    53    1.06      38      36      12      11       2       1
#> 18    8   2    53    1.06      37      37      12      12       1       1
#> 19    9   2    55    1.10      35      39      11      12       2       1
#> 20   10   2    51    1.02      39      35      14       9       3       0
#> 21    1   3    55    1.10      33      34      24       8       1       0
#> 22    2   3    58    1.16      34      32      20      12       2       0
#> 23    3   3    60    1.20      31      35      19      13       2       0
#> 24    4   3    62    1.24      29      36      20      12       3       0
#> 25    5   3    63    1.26      26      40      19      12       3       0
#> 26    6   3    60    1.20      30      38      18      10       4       0
#> 27    7   3    52    1.04      35      40      14       8       3       0
#> 28    8   3    49    0.98      37      40      13       8       2       0
#> 29    9   3    48    0.96      38      39      13       9       1       0
#> 30   10   3    49    0.98      39      37      13       9       2       0
get_nwstats(mod, sim = 2)
#>    time sim edges meandeg degree0 degree1 degree2 degree3 degree4 degree5
#> 1     1   2    55    1.10      44      23      18      10       4       1
#> 2     2   2    54    1.08      44      26      15      10       4       0
#> 3     3   2    55    1.10      44      26      14      10       5       0
#> 4     4   2    50    1.00      45      30      11       9       4       1
#> 5     5   2    49    0.98      43      34      10       9       3       1
#> 6     6   2    51    1.02      40      36      10      11       2       1
#> 7     7   2    53    1.06      38      36      12      11       2       1
#> 8     8   2    53    1.06      37      37      12      12       1       1
#> 9     9   2    55    1.10      35      39      11      12       2       1
#> 10   10   2    51    1.02      39      35      14       9       3       0
get_nwstats(mod, sim = c(1, 3))
#>    time sim edges meandeg degree0 degree1 degree2 degree3 degree4 degree5
#> 1     1   1    53    1.06      33      40      17       8       2       0
#> 2     2   1    52    1.04      35      38      17       8       2       0
#> 3     3   1    52    1.04      35      38      17       8       2       0
#> 4     4   1    53    1.06      36      37      15       9       3       0
#> 5     5   1    51    1.02      38      37      13       9       3       0
#> 6     6   1    52    1.04      37      36      16       8       3       0
#> 7     7   1    49    0.98      41      36      11       8       4       0
#> 8     8   1    50    1.00      39      37      14       6       3       1
#> 9     9   1    48    0.96      42      33      16       6       2       1
#> 10   10   1    47    0.94      44      30      18       5       2       1
#> 11    1   3    55    1.10      33      34      24       8       1       0
#> 12    2   3    58    1.16      34      32      20      12       2       0
#> 13    3   3    60    1.20      31      35      19      13       2       0
#> 14    4   3    62    1.24      29      36      20      12       3       0
#> 15    5   3    63    1.26      26      40      19      12       3       0
#> 16    6   3    60    1.20      30      38      18      10       4       0
#> 17    7   3    52    1.04      35      40      14       8       3       0
#> 18    8   3    49    0.98      37      40      13       8       2       0
#> 19    9   3    48    0.96      38      39      13       9       1       0
#> 20   10   3    49    0.98      39      37      13       9       2       0

# On the fly summary stats
summary(get_nwstats(mod))
#>       time           sim        edges          meandeg         degree0     
#>  Min.   : 1.0   Min.   :1   Min.   :47.00   Min.   :0.940   Min.   :26.00  
#>  1st Qu.: 3.0   1st Qu.:1   1st Qu.:50.00   1st Qu.:1.000   1st Qu.:35.00  
#>  Median : 5.5   Median :2   Median :52.00   Median :1.040   Median :37.50  
#>  Mean   : 5.5   Mean   :2   Mean   :52.97   Mean   :1.059   Mean   :37.37  
#>  3rd Qu.: 8.0   3rd Qu.:3   3rd Qu.:55.00   3rd Qu.:1.100   3rd Qu.:40.75  
#>  Max.   :10.0   Max.   :3   Max.   :63.00   Max.   :1.260   Max.   :45.00  
#>     degree1         degree2         degree3         degree4     
#>  Min.   :23.00   Min.   :10.00   Min.   : 5.00   Min.   :1.000  
#>  1st Qu.:34.00   1st Qu.:13.00   1st Qu.: 8.00   1st Qu.:2.000  
#>  Median :36.00   Median :14.50   Median : 9.00   Median :2.500  
#>  Mean   :35.17   Mean   :15.13   Mean   : 9.30   Mean   :2.633  
#>  3rd Qu.:38.00   3rd Qu.:17.75   3rd Qu.:10.75   3rd Qu.:3.000  
#>  Max.   :40.00   Max.   :24.00   Max.   :13.00   Max.   :5.000  
#>     degree5      
#>  Min.   :0.0000  
#>  1st Qu.:0.0000  
#>  Median :0.0000  
#>  Mean   :0.3333  
#>  3rd Qu.:1.0000  
#>  Max.   :1.0000  
colMeans(get_nwstats(mod))
#>       time        sim      edges    meandeg    degree0    degree1    degree2 
#>  5.5000000  2.0000000 52.9666667  1.0593333 37.3666667 35.1666667 15.1333333 
#>    degree3    degree4    degree5 
#>  9.3000000  2.6333333  0.3333333