Get Vertex Attribute on Network Object

Gets a vertex attribute from an object of class network. This functions simplifies the related function in the network package.

Usage

get_vertex_attribute(x, attrname)

Arguments

x

An object of class network.

attrname

The name of the attribute to get.

Value

Returns an object of class network.

Details

This function is used in EpiModel workflows to query vertex attributes on an initialized empty network object (see network_initialize).

Examples

nw <- network_initialize(100)
nw <- set_vertex_attribute(nw, "age", runif(100, 15, 65))
get_vertex_attribute(nw, "age")
#>   [1] 43.47784 59.21381 35.39877 19.44689 45.81683 41.39566 47.78289 32.67148
#>   [9] 55.99195 53.90063 49.36378 47.04044 36.00459 55.75245 42.46837 50.41658
#>  [17] 60.29835 58.51409 25.63327 16.26506 26.64246 57.27511 26.72104 49.35028
#>  [25] 21.38467 16.29437 35.25605 59.77655 30.80213 62.48113 19.90708 34.30642
#>  [33] 15.73975 20.34742 20.32534 49.05127 20.99210 27.26253 18.32362 57.01248
#>  [41] 20.14331 48.77607 28.43391 54.79549 48.95817 28.51167 39.87412 43.93837
#>  [49] 32.34548 23.59062 46.59179 30.91689 17.24170 47.09865 21.43769 59.76126
#>  [57] 60.23974 60.88306 57.70437 35.02258 57.97019 56.94412 54.93326 38.01736
#>  [65] 31.68261 30.63796 15.19070 63.49352 44.16824 39.52825 27.89115 43.93854
#>  [73] 23.61228 33.08111 34.18572 28.95392 62.07336 30.18576 42.23325 35.55849
#>  [81] 37.83716 62.77983 41.00452 37.75657 24.16395 18.76982 42.94183 48.99380
#>  [89] 35.24415 20.67075 52.87634 27.56321 41.22902 35.56261 56.83255 53.67443
#>  [97] 49.83690 29.44250 49.49479 44.52212